Update .helmignore files to minimize junk in package
When you publish a Helm chart, everything is tar'ed to a file. Ideally,
we only tar things that the templates will need to reference and nothing
else. Since we don't reference the README.md file in our templates, we
shouldn't package it either.

Each helm release will contain a copy of these files unless we ignore
them as part of the state in k8s secrets, and the helm chart git repo
will grow and grow twice as fast or so in size if we include these
files. So, that is why I care about .helmignoring them a bit.