Split apart ci workflow for the two charts
Daskhub has CRDs in it which can make us want to do things quite
differently already when validating templates and installing the charts
for basic tests, so I figure it makes sense to have them as separate
jobs and workflows as well.

By having two separate workflows, we may duplicate some CI code, but we
are enabled to run jobs based on changes to path for the specific repo
and avoid code to only do things conditionally based on the chart being
tested etc.