Improved tests and general docs/ci maintenance (#135)
* Update my own email address in Chart.yaml

* Merge daskhub's .gitignore into root folder's file

* Add dev-values.yaml for linting and tests

* Update .helmignore files to minimize junk in package

When you publish a Helm chart, everything is tar'ed to a file. Ideally,
we only tar things that the templates will need to reference and nothing
else. Since we don't reference the README.md file in our templates, we
shouldn't package it either.

Each helm release will contain a copy of these files unless we ignore
them as part of the state in k8s secrets, and the helm chart git repo
will grow and grow twice as fast or so in size if we include these
files. So, that is why I care about .helmignoring them a bit.

* Update badges etc in README.md

* Workflow revision with comments to guide readers

* Split apart ci workflow for the two charts

Daskhub has CRDs in it which can make us want to do things quite
differently already when validating templates and installing the charts
for basic tests, so I figure it makes sense to have them as separate
jobs and workflows as well.

By having two separate workflows, we may duplicate some CI code, but we
are enabled to run jobs based on changes to path for the specific repo
and avoid code to only do things conditionally based on the chart being
tested etc.

* Rename deploy to publsih-charts

* Update name from deploy to publish charts

* Add note to watch-docker-hub workflow

* Test dask chart by making an installation of it

* Update apiVersion of deprecated rbac resources

* Fix workflow syntax error

* Enable manual workflow trigger (workflow_dispatch)

* Add daskhub chart test workflow

* Badges updates again

* Make dask python package logo for python white

* Minor adjustments following PR feedback, thanks Jacob