pre-commit: remove helmlint
pre-commit.ci fails when we try to use this hook, because it is assuming
`helm` is available in the system. We have a github workflow to catch
any `helm lint` failure and we also validate the resources we create
against an old and new k8s version etc in github workflows already, so
we are very well covered.

In my experience, `helm lint` catches far less issues than `helm
template --validate` which we are running.